Implement gettimeofday and correct some functionalities.

+++ CommonFuncs.h
@@ -5,4 +5,4 @@ extern char *fileName;
 void init();
 int getRandomNum(int min, int max, int interval);
 void getFileName(char name[], int id);
-void writeOutput(FILE* fp, int totExecTime, int burstTime, int threadId);
+void writeOutput(FILE *fp, long int totExecTime, int burstTime, int threadId);
